requirements.slim.txt
attrs==23.1.0 ; python_version >= "3.8" and python_version < "3.12"
base58==2.1.1 ; python_version >= "3.8" and python_version < "3.12"
cached-property==1.5.2 ; python_version >= "3.8" and python_version < "3.12"
cattrs==23.2.2 ; python_version >= "3.8" and python_version < "3.12"
certifi==2023.11.17 ; python_version >= "3.8" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.8" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.12" and platform_system == "Windows"
cryptography==41.0.5 ; python_version >= "3.8" and python_version < "3.12"
cytoolz==0.12.2 ; python_version >= "3.8" and python_version < "3.12" and implementation_name == "cpython"
deprecation==2.1.0 ; python_version >= "3.8" and python_version < "3.12"
docker==6.1.3 ; python_version >= "3.8" and python_version < "3.12"
eth-hash==0.5.2 ; python_version >= "3.8" and python_version < "3.12"
eth-typing==3.5.2 ; python_version >= "3.8" and python_version < "3.12"
eth-utils==2.3.1 ; python_version >= "3.8" and python_version < "3.12"
exceptiongroup==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
fastecdsa==2.3.0 ; python_version >= "3.8" and python_version < "3.12"
idna==3.6 ; python_version >= "3.8" and python_version < "3.12"
importlib-resources==6.1.1 ; python_version >= "3.8" and python_version < "3.9"
jsonschema-specifications==2023.11.1 ; python_version >= "3.8" and python_version < "3.12"
jsonschema==4.20.0 ; python_version >= "3.8" and python_version < "3.12"
mnemonic==0.20 ; python_version >= "3.8" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "3.12"
netstruct==1.1.2 ; python_version >= "3.8" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.8" and python_version < "3.12"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
ply==3.11 ; python_version >= "3.8" and python_version < "3.12"
py-ecc==6.0.0 ; python_version >= "3.8" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.8" and python_version < "3.12"
pysodium==0.7.14 ; python_version >= "3.8" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.12"
pywin32==306 ; python_version >= "3.8" and python_version < "3.12" and sys_platform == "win32"
referencing==0.31.0 ; python_version >= "3.8" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.8" and python_version < "3.12"
rpds-py==0.13.1 ; python_version >= "3.8" and python_version < "3.12"
secp256k1==0.14.0 ; python_version >= "3.8" and python_version < "3.12"
setuptools==67.8.0 ; python_version >= "3.8" and python_version < "3.12"
simple-bson==0.0.3 ; python_version >= "3.8" and python_version < "3.12"
simplejson==3.19.2 ; python_version >= "3.8" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
strict-rfc3339==0.7 ; python_version >= "3.8" and python_version < "3.12"
tabulate==0.9.0 ; python_version >= "3.8" and python_version < "3.12"
testcontainers==3.7.1 ; python_version >= "3.8" and python_version < "3.12"
toolz==0.12.0 ; python_version >= "3.8" and python_version < "3.12" and (implementation_name == "pypy" or implementation_name == "cpython")
tqdm==4.66.1 ; python_version >= "3.8" and python_version < "3.12"
typing-extensions==4.8.0 ; python_version >= "3.8" and python_version < "3.12"
urllib3==2.1.0 ; python_version >= "3.8" and python_version < "3.12"
websocket-client==1.6.4 ; python_version >= "3.8" and python_version < "3.12"
wrapt==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
zipp==3.17.0 ; python_version >= "3.8" and python_version < "3.9"
